positiveInt is now represented in c# as an int (not long)

+++ src/Hl7.Fhir.Support.Tests/PrimitiveTests.cs
@@ -90,7 +90,7 @@ namespace Hl7.Fhir.Support.Tests.Model
                 { "url", typeof(string) },
                 { "string", typeof(string) },
                 { "time", typeof(PartialTime) },
-                { "positiveInt", typeof(long) },
+                { "positiveInt", typeof(int) },
                 { "integer64", typeof(long) }
             };
